tests/wpt/tests/css/css-gaps/parsing/rule-inset-cap-start-end-invalid.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title>CSS Gaps: parsing *-rule-edge-*-inset with invalid values</title>
  <link rel="author" title="Javier Contreras" href="mailto:javiercon@microsoft.com">
  <link rel="help" href="https://drafts.csswg.org/css-gaps-1/#inset">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
  <div id="target"></div>
  <script>
    // TODO(javiercon): Add shorthand support and test here.
    const properties = ['column-rule-inset-cap-start', 'row-rule-inset-cap-start', 'column-rule-inset-cap-end', 'row-rule-inset-cap-end'];
    for (const property of properties) {
      test_invalid_value(property, 'auto');
      test_invalid_value(property, 'none');
      test_invalid_value(property, '10');
      test_invalid_value(property, '10px 20px');
      test_invalid_value(property, 'overlap-join 10px');
      test_invalid_value(property, '10px overlap-join');
    }
  </script>
</body>
</html>
